+++ services/svc-api-gateway/internal/adapters/repos/devices_cache_repository_test.go
@@ -47,7 +47,7 @@ func (s *DevicesCacheRepositoryTestSuite) SetupTest() {
 
 func (s *DevicesCacheRepositoryTestSuite) TearDownTest() {
 	if s.keydbClient != nil {
-		s.keydbClient.Close()
+		_ = s.keydbClient.Close()
 	}
 	if s.miniRedis != nil {
 		s.miniRedis.Close()